 requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fontawesomefree==6.2.1
openpyxl==3.0.10
pendulum==2.1.2
pgdumplib==3.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytzdata==2020.1
rubymarshal==1.2.7
six==1.16.0
sqlparse==0.4.3
toposort==1.7
Unidecode==1.3.6
